+++ Enhancer/Plugin.cs
@@ -49,6 +49,11 @@ public class Plugin : BaseUnityPlugin
             .AddModGuidToDelegateTo("Haha.DynamicDeadline")
             .Build(),
         new PatchInfo.Builder()
+            .SetName("Death penalty")
+            .SetPatchType(typeof(DeathPenalty))
+            .SetLoadCondition(() => BoundConfig.DeathPenaltyFormulaEnabled)
+            .Build(),
+        new PatchInfo.Builder()
             .SetName("Hangar door close duration")
             .SetPatchType(typeof(HangarDoorCloseDuration))
             .SetLoadCondition(() => BoundConfig.DoorPowerDurationEnabled)
@@ -63,11 +68,12 @@ public class Plugin : BaseUnityPlugin
         new PatchInfo.Builder()
             .SetName("Item protection")
             .SetPatchType(typeof(ItemProtection))
+            .SetLoadCondition(() => BoundConfig.ScrapProtectionEnabled)
             .AddModGuidToDelegateTo("mom.llama.enhancer")
             .Build(),
         new PatchInfo.Builder()
             .SetName("Price randomizer")
-            .SetPatchType(typeof(PriceRandomizer))
+            .SetPatchType(typeof(CompanyBuyingFactorRandomizer))
             .AddModGuidToDelegateTo("mom.llama.enhancer")
             .Build(),
         new PatchInfo.Builder()