+++ Enhancer/PluginConfig.cs
@@ -4,6 +4,7 @@
     All configuration options go in here
 ***********************************************************/
 
+using BepInEx.Configuration;
 using Enhancer.Patches;
 
 namespace Enhancer;
@@ -14,19 +15,23 @@ public class PluginConfig
     public readonly bool DelegationEnabled;
     
     public readonly bool KeepConsoleEnabled;
-    public readonly bool UseRandomPrices;
     public readonly bool SuitUnlocksEnabled;
+    
+    public readonly bool RandomiseCompanyBuyingFactor;
+    public readonly float MinimumCompanyBuyingFactor;
 
     public readonly bool TimeSpeedEnabled;
     public readonly float TimeSpeed;
-    public readonly float MinimumBuyRate;
+    
     public readonly bool DoorPowerDurationEnabled;
     public readonly float DoorPowerDuration;
 
     public readonly bool StartingCreditsEnabled;
     public readonly int StartingCredits;
+    
     public readonly bool DaysPerQuotaEnabled;
     public readonly int DaysPerQuota;
+    
     public readonly bool QuotaFormulaEnabled;
     public readonly int StartingQuota;
     public readonly float QuotaIncreaseSteepness;
@@ -34,7 +39,16 @@ public class PluginConfig
     public readonly float QuotaIncreaseRandomFactor;
 
     public readonly ThreatScannerMode ThreatScanner;
-    public readonly Patches.ItemProtection.ProtectionType ScrapProtection;
+
+    public readonly bool ScrapProtectionEnabled;
+    public readonly float ScrapProtection;
+    public readonly float ScrapProtectionRandomness;
+
+    public readonly bool DeathPenaltyFormulaEnabled;
+    public readonly float MaximumDeathPenalty;
+    public readonly float MaximumDeathPenaltyPerPlayer;
+    public readonly float DeadBodyRecoveryDiscount;
+    public readonly float DeathPenaltyScalingCurvature;
 
     public PluginConfig(Plugin bindingPlugin)
     {
@@ -57,14 +71,24 @@ public class PluginConfig
             true, 
             "Whether to keep the terminal enabled after a player stops using it\nHost Required: No"
         ).Value;
-        
-        UseRandomPrices = bindingPlugin.Config.Bind(PluginInfo.PLUGIN_GUID, "bUseRandomPrices", false, "Enables the random prices setting. Great if you're using longer quota deadlines.\nThis uses a variety of things to randomize prices such as the company mood, time passed in the quota, etc.\nRespects the minimum sale value, too.\nHost Required: Yes").Value;
-        
         SuitUnlocksEnabled = bindingPlugin.Config.Bind(
             PluginInfo.PLUGIN_GUID, 
             "bUnlockSuits", 
             false, 
             "Unlocks a few of the cheaper suits from the start so your crew has options.\nHost Required: Yes"
+        ).Value;
+        
+        RandomiseCompanyBuyingFactor = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "bCompanyBuyingFactorRandomizerEnabled", 
+            false, 
+            "Randomises the company buying % when enabled. Great if you're using longer quota deadlines.\nThis uses a variety of things to randomize prices such as the company mood, time passed in the quota, etc.\nRespects the minimum sale value, too.\nHost Required: Yes"
+        ).Value;
+        MinimumCompanyBuyingFactor = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fMinimumCompanyBuyingFactor", 
+            0.0f, 
+            "The default formula for selling items to the company isn't designed to handle more than 3 days remaining.\nThe Company will be prevented from offering a factor lower than this configured value.\nRecommended values for games above 3 days: 0.3 - 0.5\nHost Required: Yes"
         ).Value;
         
         TimeSpeedEnabled = bindingPlugin.Config.Bind(
@@ -79,8 +103,6 @@ public class PluginConfig
             1.0f, 
             "How fast time passes on moons. Lower values mean time passes more slowly.\nRecommended value for single play: 0.75\nHost Required: Yes"
         ).Value;
-        
-        MinimumBuyRate = bindingPlugin.Config.Bind(PluginInfo.PLUGIN_GUID, "fMinCompanyBuyPCT", 0.0f, "The default formula for selling items to the company doesn't allow days remaining above 3.\nAlways keep this set to at least 0.0 but you probably want something higher if you have more days set for the quota.\nRecommended values for games above 3 days: 0.3 - 0.5\nHost Required: Yes").Value;
         
         DoorPowerDurationEnabled = bindingPlugin.Config.Bind(
             PluginInfo.PLUGIN_GUID,
@@ -158,6 +180,72 @@ public class PluginConfig
             ThreatScannerMode.Disabled, 
             "How the threat scanner functions. Valid types:\n - Disabled\n - Contacts: Number of Enemies on level\n - ThreatLevelPercentage: Percentage of max enemies on level\n - ThreatLevelName: Vague Text description (In order of threat level) [Clear -> Green -> Yellow -> Orange - Red]\nHost Required: No"
         ).Value;
-        ScrapProtection = bindingPlugin.Config.Bind(PluginInfo.PLUGIN_GUID, "eScrapProtection", Patches.ItemProtection.ProtectionType.SAVE_NONE, "Sets how scrap will be handled when all players die in a round.\nSAVE_NONE: Default all scrap is deleted\nSAVE_ALL: No scrap is removed\nSAVE_COINFLIP: Each piece of scrap has a 50/50 of being removed\nHost Required: Yes").Value;
+        
+        ScrapProtectionEnabled = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID,
+            "bScrapProtectionEnabled",
+            false,
+            new ConfigDescription(
+                "Sets whether or not the scrap protection feature is enabled. \nHost Required: Yes"
+            )
+        ).Value;
+        ScrapProtection = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fScrapProtection", 
+            0f, 
+            new ConfigDescription(
+                "Sets the average probability that each scrap item is kept in the event that that no players survive a mission.\nThat is, this is the approximate average fraction of secured scrap items kept.\nHost Required: Yes",
+                new AcceptableValueRange<float>(0f, 1f)
+            )
+        ).Value;
+        ScrapProtectionRandomness = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fScrapProtectionRandomnessScalar", 
+            0f, 
+            new ConfigDescription(
+                "Sets the randomness of the probability that each scrap item is kept in the event that that no players survive a mission.\n 0 -> no randomness, 0.5 -> \u00b10.5, 1 -> \u00b11\nHost Required: Yes",
+                new AcceptableValueRange<float>(0f, 1f)
+            )
+        ).Value;
+
+        DeathPenaltyFormulaEnabled = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID,
+            "bDeathPenaltyFormulaEnabled",
+            false,
+            "Feature flag for the 'death penalty formula' variables, which includes\n - 'max death penalty'\n - 'max death penalty per player'\n - 'body recovery discount'\n - 'death penalty scaling curvature'\nHost Required: Yes"
+        ).Value;
+        MaximumDeathPenalty = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fMaximumDeathPenalty", 
+            0.8f, 
+            new ConfigDescription(
+                "The maximum fraction of your money that you can lose per round.\nValue should be in [0,1], e.g.\n0 - No money can be lost.\n0.5 - Half your money can be lost in one run.\n1 - All money can be lost in one run.\nUse 0.8 for vanilla behaviour.\nHost Required: Yes",
+                new AcceptableValueRange<float>(0f, 1f)
+            )
+        ).Value;
+        MaximumDeathPenaltyPerPlayer = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fMaximumDeathPenaltyPerPlayer", 
+            0.2f, 
+            new ConfigDescription(
+                "The maximum fraction of your money that you can lose per round, per dead player.\nValue should be in [0,1].\nUse 0.2 for vanilla behaviour.\nHost Required: Yes",
+                new AcceptableValueRange<float>(0f, 1f)
+            )
+        ).Value;
+        DeadBodyRecoveryDiscount = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fDeadBodyRecoveryDiscount", 
+            0.6f, 
+            new ConfigDescription(
+                "How much recovering dead bodies reduces the penalty for that death by.\nValue should be in [0,1], e.g.\n0 - Recovering a body does not reduce the fine.\n1 - Recovering a body completely removes the fine for that death.\nUse 0.6 for vanilla behaviour.\nHost Required: Yes",
+                new AcceptableValueRange<float>(0f, 1f)
+            )
+        ).Value;
+        DeathPenaltyScalingCurvature = bindingPlugin.Config.Bind(
+            PluginInfo.PLUGIN_GUID, 
+            "fDeathPenaltyCurveDegree", 
+            0f, 
+            "How curved the death penalty scaling is. Positive -> less fine for fewer deaths. Negative -> more fine for fewer deaths.\ne.g. with a 4-player lobby:\n0 - The fine scales linearly: 25%, 50%, 75%, 100%.\n1 - The fine scales quadratically: 6.3%, 25%, 56.3%, 100%\n-1 - The fine scales anti-quadratically: 50%, 70.1%, 86.6%, 100%\nHost Required: Yes"
+        ).Value;
     }
 }